@@ -1,14 +1,15 @@
 package ru.practicum.service;
 
 import ru.practicum.EndpointHitDto;
-import ru.practicum.model.ViewStats;
+import ru.practicum.ViewStats;
+import ru.practicum.exception.InvalidDateTimeException;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 
 public interface StatService {
 
-    public String createHit(HttpServletRequest request, EndpointHitDto endpointHit);
+    public String createHit(HttpServletRequest request, EndpointHitDto endpointHit) throws InvalidDateTimeException;
 
-    public List<ViewStats> getStat(String start, String end, List<String> uris, Boolean unique);
+    public List<ViewStats> getStat(String start, String end, List<String> uris, Boolean unique) throws InvalidDateTimeException;
 }
