@@ -5,23 +5,23 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.test.annotation.DirtiesContext;
 import ru.practicum.EndpointHitDto;
+import ru.practicum.ViewStats;
+import ru.practicum.exception.InvalidDateTimeException;
 import ru.practicum.model.EndpointHit;
-import ru.practicum.model.ViewStats;
 import ru.practicum.service.StatService;
 import ru.practicum.service.StatServiceImpl;
 
-import javax.persistence.EntityManager;
-import javax.persistence.TypedQuery;
-import javax.servlet.http.HttpServletRequest;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.TypedQuery;
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 
 @SpringBootTest(
-        classes = {StatServiceImpl.class, EndpointHit.class},
+        classes = {StatServiceImpl.class, EndpointHit.class, ViewStats.class},
         properties = "db.name=test"
 )
 @EnableJpaRepositories("ru.practicum")
@@ -55,7 +55,7 @@ public class StatsServerTest {
     }
 
     @Test
-    public void saveEndpoint() {
+    public void saveEndpoint() throws InvalidDateTimeException {
         endpointHitDto.setApp("myApp1");
         service.createHit(request, endpointHitDto);
 
@@ -69,21 +69,7 @@ public class StatsServerTest {
     }
 
     @Test
-    public void saveEndpointWithNullApp() {
-        endpointHitDto.setApp(null);
-
-        Assertions.assertThrows(DataIntegrityViolationException.class, () -> service.createHit(request, endpointHitDto));
-    }
-
-    @Test
-    public void saveEndpointWithNullUri() {
-        endpointHitDto.setUri(null);
-
-        Assertions.assertThrows(DataIntegrityViolationException.class, () -> service.createHit(request, endpointHitDto));
-    }
-
-    @Test
-    public void getUnuniqueStatistic() {
+    public void getUnuniqueStatistic() throws InvalidDateTimeException {
         String start = "2020-12-31 13:27:45";
         String end = "2030-03-21 00:00:00";
         List<String> uris = new ArrayList<>();
@@ -103,7 +89,7 @@ public class StatsServerTest {
     }
 
     @Test
-    public void getUniqueStatistic() {
+    public void getUniqueStatistic() throws InvalidDateTimeException {
         String start = "2020-12-31 13:27:45";
         String end = "2030-03-21 00:00:00";
         List<String> uris = new ArrayList<>();
@@ -123,7 +109,7 @@ public class StatsServerTest {
     }
 
     @Test
-    public void getStatisticOutOfBounds() {
+    public void getStatisticOutOfBounds() throws InvalidDateTimeException {
         String start = "2027-12-31 13:27:45";
         String end = "2030-03-21 00:00:00";
         List<String> uris = new ArrayList<>();
@@ -143,7 +129,7 @@ public class StatsServerTest {
     }
 
     @Test
-    public void getStatisticWithInvalidTime() {
+    public void getStatisticWithInvalidTime() throws InvalidDateTimeException {
         String start = "2030-12-31 13:27:45";
         String end = "2028-03-21 00:00:00";
         List<String> uris = new ArrayList<>();
@@ -158,6 +144,6 @@ public class StatsServerTest {
         service.createHit(request, endpointHitDto);
         service.createHit(request, endpointHitDto2);
 
-        Assertions.assertThrows(IllegalArgumentException.class, () -> service.getStat(start, end, uris, true));
+        Assertions.assertThrows(InvalidDateTimeException.class, () -> service.getStat(start, end, uris, true));
     }
 }
