@@ -1,13 +1,17 @@
 package ru.practicum.controller;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.EndpointHitDto;
-import ru.practicum.model.ViewStats;
+import ru.practicum.ViewStats;
+import ru.practicum.exception.InvalidDateTimeException;
 import ru.practicum.service.StatServiceImpl;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
@@ -17,17 +21,20 @@ public class StatController {
 
     @Autowired
     private final StatServiceImpl service;
+    ObjectMapper mapper = new ObjectMapper();
 
+    @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/hit")
-    public String createHit(HttpServletRequest request, @RequestBody EndpointHitDto endpointHit) {
-        return service.createHit(request, endpointHit);
+    public String createHit(HttpServletRequest request, @RequestBody EndpointHitDto endpointHit) throws JsonProcessingException, InvalidDateTimeException {
+        String response = service.createHit(request, endpointHit);
+        return mapper.writeValueAsString(response);
     }
 
     @GetMapping("/stats")
     public List<ViewStats> getStat(@RequestParam(defaultValue = "2021-01-01 00:00:00") String start,
-                                   @RequestParam(defaultValue = "2021-01-01 00:00:00") String end,
+                                   @RequestParam String end,
                                    @RequestParam(required = false) List<String> uris,
-                                   @RequestParam(defaultValue = "false", required = false) Boolean unique) {
+                                   @RequestParam(defaultValue = "false", required = false) Boolean unique) throws InvalidDateTimeException {
         return service.getStat(start, end, uris, unique);
     }
 }
